COM-1917 move app file
